+++ src/oss_api.h
@@ -9,88 +9,119 @@
 
 OSS_CPP_START
 
-/**
-  * @brief  oss create bucket
-  * @param[in]  oss_acl  the oss bucket acl
-  * @return AOSE_OK success, other failure
-**/
-aos_status_t *oss_create_bucket(const oss_request_options_t *options, 
-                                const aos_string_t *bucket, 
-                                oss_acl_e oss_acl, 
+/*
+ * @brief  create oss bucket
+ * @param[in]   options       the oss request options
+ * @param[in]   bucket        the oss bucket name
+ * @param[in]   oss_acl       the oss bucket acl
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
+aos_status_t *oss_create_bucket(const oss_request_options_t *options,
+                                const aos_string_t *bucket,
+                                oss_acl_e oss_acl,
                                 aos_table_t **resp_headers);
 
-/**
-  * @brief  oss delete bucket
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  delete oss bucket
+ * @param[in]   options       the oss request options
+ * @param[in]   bucket        the oss bucket name
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_delete_bucket(const oss_request_options_t *options, 
                                 const aos_string_t *bucket, 
                                 aos_table_t **resp_headers);
 
-/**
-  * @brief  oss put bucket acl
-  * @param[out]  oss_acl  the oss bucket acl to put
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  put oss bucket acl
+ * @param[in]   options       the oss request options
+ * @param[in]   bucket        the oss bucket name
+ * @param[in]   oss_acl       the oss bucket acl
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_bucket_acl(const oss_request_options_t *options, 
                                  const aos_string_t *bucket, 
                                  oss_acl_e oss_acl,
                                  aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get bucket acl
-  * @param[out]  oss_acl  the oss bucket acl
-  * @return AOSE_OK success, other failure
-**/
-aos_status_t *oss_get_bucket_acl(const oss_request_options_t *options, 
-                                 const aos_string_t *bucket, 
-                                 aos_string_t *oss_canned_acl, 
+/*
+ * @brief  get oss bucket acl
+ * @param[in]   options       the oss request options
+ * @param[in]   bucket        the oss bucket name
+ * @param[out]  oss_acl       the oss bucket acl
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
+aos_status_t *oss_get_bucket_acl(const oss_request_options_t *options,
+                                 const aos_string_t *bucket,
+                                 aos_string_t *oss_acl,
                                  aos_table_t **resp_headers);
 
-/**
-  * @brief  oss put bucket lifycycle
-  * @param[in]  lifecycle_rule_list  the lifecycle to put
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  put oss bucket lifecycle
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   lifecycle_rule_list the oss bucket lifecycle list
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_bucket_lifecycle(const oss_request_options_t *options,
                                        const aos_string_t *bucket, 
                                        aos_list_t *lifecycle_rule_list, 
                                        aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get bucket lifecycle
-  * @param[out] lifecycle rule list
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  get oss bucket lifecycle
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[out]  lifecycle_rule_list the oss bucket lifecycle list
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_get_bucket_lifecycle(const oss_request_options_t *options,
                                        const aos_string_t *bucket, 
                                        aos_list_t *lifecycle_rule_list, 
                                        aos_table_t **resp_headers);
 
-/**
-  * @brief  oss delete bucket lifecycle
-  * @return AOSE_OK success, other failure
-  **/
+/*
+ * @brief  delete oss bucket lifecycle
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_delete_bucket_lifecycle(const oss_request_options_t *options,
                                           const aos_string_t *bucket, 
                                           aos_table_t **resp_headers);
 
-/**
-  * @brief  list object for specific bucket 
-  * @param[in]  params  parameters for list object including prefix, marker, delimiter, max_ret
-  * @param[out] params  output parameters including truncated, next_marker, obje list
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  list oss objects
+ * @param[in]   options       the oss request options
+ * @param[in]   bucket        the oss bucket name
+ * @param[in]   params        input params for list object request,
+                              including prefix, marker, delimiter, max_ret
+ * @param[out]  params        output params for list object response,
+                              including truncated, next_marker, obje list
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_list_object(const oss_request_options_t *options, 
                               const aos_string_t *bucket, 
                               oss_list_object_params_t *params, 
                               aos_table_t **resp_headers);
 
-/**
-  * @brief  oss put object from buffer
-  * @param[in]  buffer  the buffer containing object content
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  put oss object from buffer
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   buffer              the buffer containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_object_from_buffer(const oss_request_options_t *options, 
                                          const aos_string_t *bucket, 
                                          const aos_string_t *object, 
@@ -98,11 +129,16 @@ aos_status_t *oss_put_object_from_buffer(const oss_request_options_t *options,
                                          aos_table_t *headers,
                                          aos_table_t **resp_headers);
 
-/**
-  * @brief  oss put object from file
-  * @param[in]  filename  the filename containing object content  
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  put oss object from file
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   filename            the filename to put
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_object_from_file(const oss_request_options_t *options,
                                        const aos_string_t *bucket, 
                                        const aos_string_t *object, 
@@ -110,11 +146,17 @@ aos_status_t *oss_put_object_from_file(const oss_request_options_t *options,
                                        aos_table_t *headers, 
                                        aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get object content to buffer
-  * @param[out]  buffer  the buffer containing download object content
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  get oss object to buffer
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   headers             the headers for request
+ * @param[in]   params              the params for request
+ * @param[out]  buffer              the buffer containing object content
+ * @param[out]  resp_headers  oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_get_object_to_buffer(const oss_request_options_t *options, 
                                        const aos_string_t *bucket, 
                                        const aos_string_t *object,
@@ -123,11 +165,17 @@ aos_status_t *oss_get_object_to_buffer(const oss_request_options_t *options,
                                        aos_list_t *buffer, 
                                        aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get object content to file
-  * @param[in] filename  the file containing download object content
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  get oss object to file
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   headers             the headers for request
+ * @param[in]   params              the params for request
+ * @param[out]  filename            the filename storing object content
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_get_object_to_file(const oss_request_options_t *options,
                                      const aos_string_t *bucket, 
                                      const aos_string_t *object,
@@ -136,49 +184,73 @@ aos_status_t *oss_get_object_to_file(const oss_request_options_t *options,
                                      aos_string_t *filename, 
                                      aos_table_t **resp_headers);
 
-/**
-  * @brief  oss head object
-  * @param[out]  resp_headers  the response headers containing object meta
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  head oss object
+ * @param[in]   options          the oss request options
+ * @param[in]   bucket           the oss bucket name
+ * @param[in]   object           the oss object name
+ * @param[in]   headers          the headers for request
+ * @param[out]  resp_headers     oss server response headers containing object meta
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_head_object(const oss_request_options_t *options, 
                               const aos_string_t *bucket, 
                               const aos_string_t *object,
                               aos_table_t *headers, 
                               aos_table_t **resp_headers);
 
-/**
-  * @brief  oss delete object
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  delete oss object
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_delete_object(const oss_request_options_t *options, 
                                 const aos_string_t *bucket, 
                                 const aos_string_t *object, 
                                 aos_table_t **resp_headers);
 
-/**
-  * @brief  oss delete objects
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  delete oss objects
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object_list         the oss object list name
+ * @param[in]   is_quiet            is quiet or verbose
+ * @param[out]  resp_headers        oss server response headers
+ * @param[out]  deleted_object_list deleted object list
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_delete_objects(const oss_request_options_t *options,
                                  const aos_string_t *bucket, 
-                                 aos_list_t *object_list, int is_quiet,
+                                 aos_list_t *object_list, 
+                                 int is_quiet,
                                  aos_table_t **resp_headers, 
                                  aos_list_t *deleted_object_list);
 
-/**
-  * @brief  oss delete objects by prefix
-  * @param[in]  prefix  prefix of delete objects
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  delete oss objects by prefix
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   prefix              prefix of delete objects
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_delete_objects_by_prefix(oss_request_options_t *options,
                                            const aos_string_t *bucket, 
                                            const aos_string_t *prefix);
 
-/**
-  * @brief  oss copy object from source object
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  copy oss objects
+ * @param[in]   options             the oss request options
+ * @param[in]   source_bucket       the oss source bucket name
+ * @param[in]   object_list         the oss source object list name
+ * @param[in]   dest_bucket         the oss dest bucket name
+ * @param[in]   dest_list           the oss dest object list name
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_copy_object(const oss_request_options_t *options, 
                               const aos_string_t *source_bucket, 
                               const aos_string_t *source_object, 
@@ -193,6 +265,18 @@ aos_status_t *oss_copy_object(const oss_request_options_t *options,
   * @param[in]  buffer  the buffer containing appending content  
   * @return AOSE_OK success, other failure
 **/
+
+/*
+ * @brief  append oss object from buffer
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   position            the start position append
+ * @param[in]   buffer              the buffer containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_append_object_from_buffer(const oss_request_options_t *options,
                                             const aos_string_t *bucket, 
                                             const aos_string_t *object, 
@@ -201,12 +285,17 @@ aos_status_t *oss_append_object_from_buffer(const oss_request_options_t *options
                                             aos_table_t *headers, 
                                             aos_table_t **resp_headers);
 
-/**
-  * @brief  oss append object from file
-  * @param[in]  position  the start position append after object
-  * @param[in]  append_file  the file containing appending content 
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  append oss object from file
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   position            the start position append
+ * @param[in]   append_file         the file containing appending content 
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_append_object_from_file(const oss_request_options_t *options,
                                           const aos_string_t *bucket, 
                                           const aos_string_t *object, 
@@ -221,90 +310,123 @@ aos_status_t *oss_append_object_from_file(const oss_request_options_t *options,
   * @param[in]  expires  the expire time in Unix timestamp format
   * @return the signed url 
 **/
+
+/*
+ * @brief  gen signed url for oss object api
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   expires             the end expire time for signed url
+ * @param[in]   req                 the aos http request
+ * @return  signed url, non-NULL success, NULL failure
+ */
 char *oss_gen_signed_url(const oss_request_options_t *options, 
                          const aos_string_t *bucket,
                          const aos_string_t *object, 
                          int64_t expires, 
                          aos_http_request_t *req);
 
-/**
-  * @brief  oss put object from buffer using signed url
-  * @param[in]  signed_url  the signed url for object
-  * @param[in]  buffer  the buffer containing object content
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss put object from buffer using signed url
+ * @param[in]   options             the oss request options
+ * @param[in]   signed_url          the signed url for put object
+ * @param[in]   buffer              the buffer containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_object_from_buffer_by_url(const oss_request_options_t *options,
                                                 const aos_string_t *signed_url, 
                                                 aos_list_t *buffer, 
                                                 aos_table_t *headers,
                                                 aos_table_t **resp_headers);
 
-/**
-  * @brief  oss put object from file using signed_url
-  * @param[in]  signed_url  the signed url for object
-  * @param[in]  filename  the file containing object content  
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss put object from file using signed url
+ * @param[in]   options             the oss request options
+ * @param[in]   signed_url          the signed url for put object
+ * @param[in]   filename            the filename containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_put_object_from_file_by_url(const oss_request_options_t *options,
                                               const aos_string_t *signed_url, 
                                               aos_string_t *filename, 
                                               aos_table_t *headers,
                                               aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get object to buffer using signed url
-  * @param[in]  signed_url  the signed url for object
-  * @param[in]  buffer  the buffer containing download object content 
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss get object to buffer using signed url
+ * @param[in]   options             the oss request options
+ * @param[in]   signed_url          the signed url for put object
+ * @param[in]   buffer              the buffer containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_get_object_to_buffer_by_url(const oss_request_options_t *options,
                                               const aos_string_t *signed_url, 
-                                              aos_table_t *headers, 
                                               aos_list_t *buffer, 
+                                              aos_table_t *headers,
                                               aos_table_t **resp_headers);
 
-/**
-  * @brief  oss get object to file using signed url
-  * @param[in]  signed_url  the signed url for object
-  * @param[in]  filename  the file containing download object content 
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss get object to file using signed url
+ * @param[in]   options             the oss request options
+ * @param[in]   signed_url          the signed url for put object
+ * @param[in]   filename            the filename containing object content
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_get_object_to_file_by_url(const oss_request_options_t *options,
                                             const aos_string_t *signed_url, 
                                             aos_table_t *headers, 
                                             aos_string_t *filename,
                                             aos_table_t **resp_headers);
 
-/**
-  * @brief  oss head object using signed url
-  * @param[in]  signed_url  the signed url for object
-  * @param[out]  resp_headers  the response headers containing object meta
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss head object using signed url
+ * @param[in]   options             the oss request options
+ * @param[in]   signed_url          the signed url for put object
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_head_object_by_url(const oss_request_options_t *options,
                                      const aos_string_t *signed_url, 
                                      aos_table_t *headers, 
                                      aos_table_t **resp_headers);
 
-/**
-  * @brief  oss init multipart upload
-  * @param[out] upload_id  the returned upload id 
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss init multipart upload
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_init_multipart_upload(const oss_request_options_t *options, 
                                         const aos_string_t *bucket, 
                                         const aos_string_t *object, 
-                                        aos_table_t *headers, 
                                         aos_string_t *upload_id,
+                                        aos_table_t *headers,
                                         aos_table_t **resp_headers);
 
-/**
-  * @brief  oss upload part from buffer
-  * @param[in]  upload_id  the upload id
-  * @param[in]  part_num  the upload part number
-  * @param[in]  buffer  the buffer containing upload part content 
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss upload part from buffer
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   part_num            the upload part number
+ * @param[in]   buffer              the buffer containing upload part content
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_upload_part_from_buffer(const oss_request_options_t *options, 
                                           const aos_string_t *bucket, 
                                           const aos_string_t *object, 
@@ -313,13 +435,17 @@ aos_status_t *oss_upload_part_from_buffer(const oss_request_options_t *options,
                                           aos_list_t *buffer, 
                                           aos_table_t **resp_headers);
 
-/**
-  * @brief  oss upload part from file
-  * @param[in]  upload_id  the upload id
-  * @param[in]  part_num  the upload part number
-  * @param[in]  upload_file  the file params for range upload
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss upload part from file
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   part_num            the upload part number
+ * @param[in]   upload_file         the file containing upload part content
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_upload_part_from_file(const oss_request_options_t *options,
                                         const aos_string_t *bucket, 
                                         const aos_string_t *object,
@@ -328,23 +454,33 @@ aos_status_t *oss_upload_part_from_file(const oss_request_options_t *options,
                                         oss_upload_file_t *upload_file,
                                         aos_table_t **resp_headers);
 
-/**
-  * @brief  oss abort multipart upload
-  * @param[in]  upload_id  the upload id
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss abort multipart upload
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_abort_multipart_upload(const oss_request_options_t *options, 
                                          const aos_string_t *bucket, 
                                          const aos_string_t *object, 
                                          aos_string_t *upload_id, 
                                          aos_table_t **resp_headers);
 
-/**
-  * @brief  oss complete multipart upload
-  * @param[in]  upload_id  the upload id
-  * @param[in]  part_list  the uploaded part list to complete
-  * @return AOSE_OK success, other failure
-**/
+
+/*
+ * @brief  oss complete multipart upload
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   part_list           the uploaded part list to complete
+ * @param[in]   headers             the headers for request          
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_complete_multipart_upload(const oss_request_options_t *options, 
                                             const aos_string_t *bucket, 
                                             const aos_string_t *object, 
@@ -353,13 +489,19 @@ aos_status_t *oss_complete_multipart_upload(const oss_request_options_t *options
                                             aos_table_t *headers,
                                             aos_table_t **resp_headers);
 
-/**
-  * @brief  oss list upload part with specific upload_id for object
-  * @param[in]  upload_id  the upload id
-  * @param[in]  params  the list upload part parameters incluing part_number_marker, max_ret
-  * @param[out]  params  the output params including next_part_number_marker, part_list, truncated
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss list upload part with specific upload_id for object
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   params              the input list upload part parameters,
+                                    incluing part_number_marker, max_ret
+ * @param[out]  params              the output params,
+                                    including next_part_number_marker, part_list, truncated
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_list_upload_part(const oss_request_options_t *options, 
                                    const aos_string_t *bucket, 
                                    const aos_string_t *object, 
@@ -367,32 +509,44 @@ aos_status_t *oss_list_upload_part(const oss_request_options_t *options,
                                    oss_list_upload_part_params_t *params, 
                                    aos_table_t **resp_headers);
 
-/**
-  * @brief  oss list multipart upload for bucket
-  * @param[in]  params  the list multipart upload parameters
-  * @param[out]  params  the output params including next_key_marker, next_upload_id_markert, upload_list etc
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss list multipart upload for bucket
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   params              the input list multipart upload parameters
+ * @param[out]  params              the output params including next_key_marker, next_upload_id_markert, upload_list etc
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_list_multipart_upload(const oss_request_options_t *options, 
                                         const aos_string_t *bucket, 
                                         oss_list_multipart_upload_params_t *params, 
                                         aos_table_t **resp_headers);
 
-/**
-  * @brief  oss copy large object using upload part copy
-  * @param[in]  params  the upload part copy parameters
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss copy large object using upload part copy
+ * @param[in]   options             the oss request options
+ * @param[in]   paramsthe           upload part copy parameters
+ * @param[in]   headers             the headers for request
+ * @param[out]  resp_headers        oss server response headers
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_upload_part_copy(const oss_request_options_t *options,
                                    oss_upload_part_copy_params_t *params, 
                                    aos_table_t *headers, 
                                    aos_table_t **resp_headers);
 
-/**
-  * @brief  oss upload file using multipart upload
-  * @param[in]  part_size  the part size for multipart upload
-  * @return AOSE_OK success, other failure
-**/
+/*
+ * @brief  oss upload file using multipart upload
+ * @param[in]   options             the oss request options
+ * @param[in]   bucket              the oss bucket name
+ * @param[in]   object              the oss object name
+ * @param[in]   upload_id           the upload id to upload if has
+ * @param[in]   filename            the filename containing object content
+ * @param[in]   part_size           the part size for multipart upload
+ * @param[in]   headers             the headers for request
+ * @return  aos_status_t, code is 2xx success, other failure
+ */
 aos_status_t *oss_upload_file(oss_request_options_t *options,
                               const aos_string_t *bucket, 
                               const aos_string_t *object,